Don’t attempt to purge cache for invalid titles
For pages which are in an entity namespace, but whose title doesn’t form
a valid entity ID, the content model is still an entity content model
but EntityContentFactory::getEntityIdForTitle will return null instead
of an entity ID. ResultsCache::delete can’t be called with a null
argument, so skip the purging in this case (there’s nothing to purge
anyways).

Bug: T191413

+++ src/WikibaseQualityConstraintsHooks.php
@@ -76,8 +76,10 @@ final class WikibaseQualityConstraintsHooks {
 		$entityContentFactory = $repo->getEntityContentFactory();
 		if ( $entityContentFactory->isEntityContentModel( $wikiPage->getContentModel() ) ) {
 			$entityId = $entityContentFactory->getEntityIdForTitle( $wikiPage->getTitle() );
-			$resultsCache = ResultsCache::getDefaultInstance();
-			$resultsCache->delete( $entityId );
+			if ( $entityId !== null ) {
+				$resultsCache = ResultsCache::getDefaultInstance();
+				$resultsCache->delete( $entityId );
+			}
 		}
 	}
